FastAPI/requirements.txt
alembic==1.7.7
aiofiles==0.5.0
bcrypt==3.2.0
cffi==1.15.0
click==7.1.2
cryptography==35.0.0
ecdsa==0.17.0
fastapi==0.68.1
greenlet==1.1.2
h11==0.13.0
Jinja2==3.0.3
MarkupSafe==2.1.0
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0

python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
typing_extensions==4.1.1
uvicorn==0.13.4
